Extract BindingGraphFactory from BindingGraph.
RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=185899992
